Review comment fixes
