+++ clients/spark-extensions/v3.3/src/test/java/org/projectnessie/spark/extensions/ITNessieStatements.java
@@ -15,17 +15,4 @@
  */
 package org.projectnessie.spark.extensions;
 
-import org.apache.iceberg.spark.extensions.IcebergSparkSessionExtensions;
-import org.junit.jupiter.api.BeforeAll;
-
-public class ITNessieStatements extends AbstractNessieSparkSqlExtensionTest {
-
-  @BeforeAll
-  protected static void useNessieExtensions() {
-    conf.set(
-        "spark.sql.extensions",
-        NessieSparkSessionExtensions.class.getCanonicalName()
-            + ","
-            + IcebergSparkSessionExtensions.class.getCanonicalName());
-  }
-}
+public class ITNessieStatements extends AbstractNessieSparkSqlExtensionTest {}
